# ARROW-11030: [Rust][DataFusion] Concatenate left side batches to single batch in HashJoinExec