This PR removes the exponential slowdown based on the number of left/right batches in the hash join by concatenating all left batches to a single batch, so we can directly index into the array (with `take`) instead of building a structure / iterating all the left arrays. This also simplifies the code, which can help with further speed improvements and reducing memory usage.

The benchmark results look very promising, mostly removing the overhead of small batches and big datasets with many left-side batches. For a very small batch size of 64, on tcph q 12:

```
Query 12 iteration 0 took 1596.1 ms
Query 12 iteration 1 took 1602.6 ms
Query 12 iteration 2 took 1594.3 ms
Query 12 iteration 3 took 1610.6 ms
Query 12 iteration 4 took 1615.9 ms
Query 12 iteration 5 took 1619.8 ms
Query 12 iteration 6 took 1645.5 ms
Query 12 iteration 7 took 1653.7 ms
Query 12 iteration 8 took 1644.0 ms
Query 12 iteration 9 took 1641.9 ms
Query 12 avg time: 1622.43 ms
```

Master

```
Query 12 iteration 0 took 10701.6 ms
Query 12 iteration 1 took 10949.0 ms
Query 12 iteration 2 took 11556.2 ms
Query 12 iteration 3 took 11929.1 ms
Query 12 iteration 4 took 11850.4 ms
Query 12 iteration 5 took 11993.6 ms
Query 12 iteration 6 took 12132.9 ms
Query 12 iteration 7 took 12637.8 ms
Query 12 iteration 8 took 12561.6 ms
Query 12 iteration 9 took 12887.9 ms
Query 12 avg time: 11920.01 ms
```


The PR also includes improvements on query 1 / 5 (batch size 64k):

This PR:
```
Query 1 iteration 0 took 571.0 ms
Query 1 iteration 1 took 566.0 ms
Query 1 iteration 2 took 569.0 ms
Query 1 iteration 3 took 565.1 ms
Query 1 iteration 4 took 569.6 ms
Query 1 iteration 5 took 567.7 ms
Query 1 iteration 6 took 567.7 ms
Query 1 iteration 7 took 564.9 ms
Query 1 iteration 8 took 566.2 ms
Query 1 iteration 9 took 567.1 ms
Query 1 avg time: 567.44 ms
```

Master

```
Query 1 iteration 0 took 647.8 ms
Query 1 iteration 1 took 632.8 ms
Query 1 iteration 2 took 634.1 ms
Query 1 iteration 3 took 629.0 ms
Query 1 iteration 4 took 631.8 ms
Query 1 iteration 5 took 639.6 ms
Query 1 iteration 6 took 640.1 ms
Query 1 iteration 7 took 628.1 ms
Query 1 iteration 8 took 635.3 ms
Query 1 iteration 9 took 642.9 ms
Query 1 avg time: 636.16 ms
```

PR
```
Query 5 iteration 0 took 628.6 ms
Query 5 iteration 1 took 604.9 ms
Query 5 iteration 2 took 619.2 ms
Query 5 iteration 3 took 614.4 ms
Query 5 iteration 4 took 629.2 ms
Query 5 iteration 5 took 617.0 ms
Query 5 iteration 6 took 618.9 ms
Query 5 iteration 7 took 636.1 ms
Query 5 iteration 8 took 639.5 ms
Query 5 iteration 9 took 636.1 ms
Query 5 avg time: 624.39 ms
```

Master:
```
Query 5 iteration 0 took 662.2 ms
Query 5 iteration 1 took 715.7 ms
Query 5 iteration 2 took 671.4 ms
Query 5 iteration 3 took 664.9 ms
Query 5 iteration 4 took 645.6 ms
Query 5 iteration 5 took 655.9 ms
Query 5 iteration 6 took 657.7 ms
Query 5 iteration 7 took 634.3 ms
Query 5 iteration 8 took 658.3 ms
Query 5 iteration 9 took 665.6 ms
Query 5 avg time: 663.15 ms
```