@andygrove 2021-01-01
I'm seeing a 2x improvement in performance for q12 @ SF=100 with batch size 4096 :rocket: 

This should allow us to reduce the default batch size and reduce memory pressure.