@Dandandan 2021-01-03
@andygrove @jorgecarleitao 

Is now ready for review. 
I added some comments myself too.

The PR now includes the following:

* The main change to collect into 1 batch, and use `take` directly
* Build info about column indexes upfront
* Speed improvement in `create_key` (doing `row.value(i)` twice most importantly), this also makes q1 and q5 run a bit faster
* Misc refactoring