@andygrove 2021-01-03
Yes @jorgecarleitao reducing the number of batches makes sense here. I am just slightly concerned that always reducing to a single batch may be introducing a constraint that could bit us in the future, although this is realistically only an issue when using 32-bit offsets for variable-width arrays.