@Dandandan 2021-01-05
@jorgecarleitao @andygrove 
Do you think this design is good enough for now?
For now it means anyone storing > 4GB variable size data in a build-side column should opt to use 64-bit offsets, and in the future one option would be to automatically detect this and apply a `cast` when needed. I think that should have the least performance implications.